plant_detector/requirements.txt
Flask
opencv-python
numpy
requests
python-dotenv
Pillow>=10.0.0
google-generativeai>=0.6.0
